+++ Domain/StaffApp.Domain/Entity/UserAppraisal.cs
@@ -9,7 +9,7 @@ namespace StaffApp.Domain.Entity
         public int AppraisalPeriodId { get; set; }
         public string UserId { get; set; }
         public string? ReviewerId { get; set; }
-        public decimal OverallRating { get; set; }
+        public double OverallRating { get; set; }
         public string Comments { get; set; }
         public string AreaForDevelopment { get; set; }
         public string GoalsForNextPeriod { get; set; }
